requirements.txt
# Lista de dependências para o chatbot da Laura
streamlit>=1.23.0
pypdf>=3.17.0
faiss-cpu>=1.7.4
openai>=1.0.0
tiktoken>=0.5.1